+++ src/components/world/AutoColorizedImage.tsx
@@ -1,31 +1,28 @@
 import React, { useCallback, useEffect, useRef } from "react";
-import { Palette } from "shared/lib/entities/entitiesTypes";
-import ColorizedImageWorker, {
-  ColorizedImageResult,
-} from "./ColorizedImage.worker";
+import AutoColorizedImageWorker, {
+  AutoColorizedImageResult,
+} from "./AutoColorizedImage.worker";
 
-const workerPool: ColorizedImageWorker[] = [];
+const workerPool: AutoColorizedImageWorker[] = [];
 for (let i = 0; i < navigator.hardwareConcurrency; i++) {
-  workerPool.push(new ColorizedImageWorker());
+  workerPool.push(new AutoColorizedImageWorker());
 }
 
-interface ColorizedImageProps {
+interface AutoColorizedImageProps {
   width: number;
   height: number;
   src: string;
-  tiles: number[];
-  palettes: Palette[];
+  tilesSrc?: string;
   previewAsMono?: boolean;
 }
 
-const ColorizedImage = ({
+const AutoColorizedImage = ({
   width,
   height,
   src,
-  tiles,
-  palettes,
+  tilesSrc,
   previewAsMono,
-}: ColorizedImageProps) => {
+}: AutoColorizedImageProps) => {
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const workerId = useRef(Math.random());
   const worker = useRef(
@@ -33,7 +30,7 @@ const ColorizedImage = ({
   );
 
   const onWorkerComplete = useCallback(
-    (e: MessageEvent<ColorizedImageResult>) => {
+    (e: MessageEvent<AutoColorizedImageResult>) => {
       if (e.data.id === workerId.current) {
         const offscreenCanvas = document.createElement("canvas");
         const offscreenCtx = offscreenCanvas.getContext("bitmaprenderer");
@@ -64,17 +61,16 @@ const ColorizedImage = ({
     if (canvasRef.current && worker.current) {
       worker.current.postMessage({
         src,
-        palettes: palettes.map((p) => p.colors),
-        tiles,
+        tilesSrc,
         width,
         height,
         previewAsMono,
         id: workerId.current,
       });
     }
-  }, [height, palettes, previewAsMono, src, tiles, width]);
+  }, [height, src, tilesSrc, width, previewAsMono]);
 
   return <canvas ref={canvasRef} width={width} height={height} />;
 };
 
-export default ColorizedImage;
+export default AutoColorizedImage;