+++ src/lib/helpers/fs/fileExists.ts
@@ -1,0 +1,10 @@
+import { promises as fs } from "fs";
+
+export const fileExists = async (filePath: string): Promise<boolean> => {
+  try {
+    await fs.access(filePath);
+    return true;
+  } catch (error) {
+    return false;
+  }
+};