+++ src/store/configureStore.ts
@@ -12,6 +12,7 @@ import spriteMiddleware from "./features/sprite/spriteMiddleware";
 import throttleMiddleware from "./features/throttle/throttleMiddleware";
 import trackerDocumentMiddleware from "./features/trackerDocument/trackerDocumentMiddleware";
 import entitiesMiddleware from "./features/entities/entitiesMiddleware";
+import settingsMiddleware from "./features/settings/settingsMiddleware";
 
 export type RootState = ReturnType<typeof rootReducer>;
 
@@ -29,6 +30,7 @@ const store = configureStore({
     electronMiddleware,
     projectMiddleware,
     entitiesMiddleware,
+    settingsMiddleware,
     spriteMiddleware,
     buildGameMiddleware,
     musicMiddleware,